implement FromStr for *_parse cases
Fixes #663
