+++ gdk4/src/content_formats.rs
@@ -28,3 +28,13 @@ impl ContentFormats {
         }
     }
 }
+
+#[cfg(any(feature = "v4_4", feature = "dox"))]
+#[cfg_attr(feature = "dox", doc(cfg(feature = "v4_4")))]
+impl std::str::FromStr for ContentFormats {
+    type Err = glib::BoolError;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        skip_assert_initialized!();
+        ContentFormats::parse(s)
+    }
+}
