+++ gsk4/Gir.toml
@@ -153,3 +153,6 @@ status = "generate"
     [[object.function]]
     name = "print"
     ignore = true # Same as Transform::to_string
+    [[object.function]]
+    name = "parse"
+    manual = true # bool_return_is_error doesn't seem to work for it
